+++ supabase/schemas/05_rls/05-03_conversations_rls.sql
@@ -31,8 +31,9 @@ using (
   -- 04-02_visibility_helpers.sql.
   and (
     (
-      (organization_id, organization_address) in (
-        select v.organization_id, v.address from public.get_visible_addresses() v
+      (organization_id, service, organization_address) in (
+        select v.organization_id, v.service, v.address
+        from public.get_visible_addresses() v
       )
       and id not in (select public.get_restricted_conversations())
     )
@@ -70,8 +71,9 @@ using (
   and service = 'local'::public.service
   and (
     (
-      (organization_id, organization_address) in (
-        select v.organization_id, v.address from public.get_visible_addresses() v
+      (organization_id, service, organization_address) in (
+        select v.organization_id, v.service, v.address
+        from public.get_visible_addresses() v
       )
       and id not in (select public.get_restricted_conversations())
     )
@@ -96,8 +98,9 @@ using (
   and service = 'local'::public.service
   and (
     (
-      (organization_id, organization_address) in (
-        select v.organization_id, v.address from public.get_visible_addresses() v
+      (organization_id, service, organization_address) in (
+        select v.organization_id, v.service, v.address
+        from public.get_visible_addresses() v
       )
       and id not in (select public.get_restricted_conversations())
     )